requirements.txt
asgiref
Django
django-extensions
djangorestframework
pillow
sqlparse
tzdata
django-silk

# Added for Postgres and .env support
python-dotenv
psycopg2-binary
